refactoring of sourcefile classes
